refactor(platform): split platform modules
